Moved everything to `parsers` directory
