+++ packages/shared/src/lib/utils.shared.ts
@@ -2,6 +2,7 @@ import {v4 as uuidv4} from 'uuid';
 
 import type {AsyncResult} from '@shared/types/result.types';
 import {makeErrorResult, makeSuccessResult} from '@shared/types/result.types';
+import type {EmailAddress} from '@shared/types/user.types';
 import type {Func, Supplier, UUID} from '@shared/types/utils.types';
 
 /**
@@ -84,4 +85,13 @@ export function partition<T, U>(arr: Array<T | U>, predicate: Func<T | U, boolea
  */
 export function makeUuid<T = UUID>(): T {
   return uuidv4() as T;
+}
+
+const EMAIL_REGEX = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+
+/**
+ * Checks if a value is a valid `EmailAddress`.
+ */
+export function isValidEmail(maybeEmail: unknown): maybeEmail is EmailAddress {
+  return typeof maybeEmail === 'string' && EMAIL_REGEX.test(maybeEmail);
 }
