+++ packages/shared/src/types/feedItems.types.ts
@@ -6,7 +6,8 @@ import {makeUuid} from '@shared/lib/utils.shared';
 import type {IconName} from '@shared/types/icons.types';
 import type {KeyboardShortcutId} from '@shared/types/shortcuts.types';
 import type {TagId} from '@shared/types/tags.types';
-import {UserIdSchema, type UserId} from '@shared/types/user.types';
+import type {UserId} from '@shared/types/user.types';
+import {UserIdSchema} from '@shared/types/user.types';
 import type {UserFeedSubscriptionId} from '@shared/types/userFeedSubscriptions.types';
 import {UserFeedSubscriptionIdSchema} from '@shared/types/userFeedSubscriptions.types';
 import type {BaseStoreItem, Timestamp} from '@shared/types/utils.types';
