+++ packages/shared/src/types/importQueue.types.ts
@@ -1,19 +1,12 @@
 import {z} from 'zod';
 
-import {
-  parseZodResult,
-  prefixErrorResult,
-  prefixResultIfError,
-} from '@shared/lib/errorUtils.shared';
 import {makeUuid} from '@shared/lib/utils.shared';
-
-import {parseFeedItemId} from '@shared/parsers/feedItems.parser';
 
 import {FeedItemIdSchema} from '@shared/types/feedItems.types';
 import type {FeedItemId} from '@shared/types/feedItems.types';
 import type {Result} from '@shared/types/result.types';
 import {makeSuccessResult} from '@shared/types/result.types';
-import {parseUserId, UserIdSchema} from '@shared/types/user.types';
+import {UserIdSchema} from '@shared/types/user.types';
 import type {UserId} from '@shared/types/user.types';
 import type {BaseStoreItem, Timestamp} from '@shared/types/utils.types';
 
@@ -24,18 +17,6 @@ import type {BaseStoreItem, Timestamp} from '@shared/types/utils.types';
 export type ImportQueueItemId = string & {readonly __brand: 'ImportQueueItemIdBrand'};
 
 export const ImportQueueItemIdSchema = z.string().uuid();
-
-/**
- * Parses a {@link ImportQueueItemId} from a plain string. Returns an `ErrorResult` if the string is
- * not valid.
- */
-export function parseImportQueueItemId(maybeImportQueueItemId: string): Result<ImportQueueItemId> {
-  const parsedResult = parseZodResult(ImportQueueItemIdSchema, maybeImportQueueItemId);
-  if (!parsedResult.success) {
-    return prefixErrorResult(parsedResult, 'Invalid import queue item ID');
-  }
-  return makeSuccessResult(parsedResult.value as ImportQueueItemId);
-}
 
 /**
  * Creates a new random {@link ImportQueueItemId}.
@@ -72,39 +53,6 @@ export const ImportQueueItemSchema = z.object({
   createdTime: z.date(),
   lastUpdatedTime: z.date(),
 });
-
-/**
- * Parses a {@link FeedSource} from an unknown value. Returns an `ErrorResult` if the value is not
- * valid.
- */
-export function parseImportQueueItem(maybeImportQueueItem: unknown): Result<ImportQueueItem> {
-  const parsedImportQueueItemResult = parseZodResult(ImportQueueItemSchema, maybeImportQueueItem);
-  if (!parsedImportQueueItemResult.success) {
-    return prefixResultIfError(parsedImportQueueItemResult, 'Invalid import queue item');
-  }
-
-  const parsedImportQueueItemIdResult = parseImportQueueItemId(
-    parsedImportQueueItemResult.value.importQueueItemId
-  );
-  if (!parsedImportQueueItemIdResult.success) return parsedImportQueueItemIdResult;
-
-  const parsedUserIdResult = parseUserId(parsedImportQueueItemResult.value.userId);
-  if (!parsedUserIdResult.success) return parsedUserIdResult;
-
-  const parsedFeedItemIdResult = parseFeedItemId(parsedImportQueueItemResult.value.feedItemId);
-  if (!parsedFeedItemIdResult.success) return parsedFeedItemIdResult;
-
-  const {url, createdTime, lastUpdatedTime} = parsedImportQueueItemResult.value;
-  return makeSuccessResult({
-    importQueueItemId: parsedImportQueueItemIdResult.value,
-    userId: parsedUserIdResult.value,
-    feedItemId: parsedFeedItemIdResult.value,
-    url,
-    status: parsedImportQueueItemResult.value.status,
-    createdTime: new Date(createdTime) as unknown as Timestamp,
-    lastUpdatedTime: new Date(lastUpdatedTime) as unknown as Timestamp,
-  });
-}
 
 /**
  * An item in the feed import queue.
