+++ packages/shared/src/types/userFeedSubscriptions.types.ts
@@ -1,17 +1,16 @@
 import {z} from 'zod';
 
-import {parseZodResult, prefixErrorResult} from '@shared/lib/errorUtils.shared';
 import {makeUuid} from '@shared/lib/utils.shared';
 
 import {
   FeedSourceIdSchema,
-  parseFeedSourceId,
   type FeedSource,
   type FeedSourceId,
 } from '@shared/types/feedSources.types';
 import type {Result} from '@shared/types/result.types';
 import {makeSuccessResult} from '@shared/types/result.types';
-import {parseUserId, UserIdSchema, type UserId} from '@shared/types/user.types';
+import type {UserId} from '@shared/types/user.types';
+import {UserIdSchema} from '@shared/types/user.types';
 import type {BaseStoreItem, Timestamp} from '@shared/types/utils.types';
 
 /**
@@ -54,63 +53,10 @@ export interface UserFeedSubscription extends BaseStoreItem {
 }
 
 /**
- * Parses a {@link UserFeedSubscriptionId} from a plain string. Returns an `ErrorResult` if the
- * string is not valid.
- */
-export function parseUserFeedSubscriptionId(
-  maybeUserFeedSubscriptionId: string
-): Result<UserFeedSubscriptionId> {
-  const parsedResult = parseZodResult(UserFeedSubscriptionIdSchema, maybeUserFeedSubscriptionId);
-  if (!parsedResult.success) {
-    return prefixErrorResult(parsedResult, 'Invalid user feed subscription ID');
-  }
-  return makeSuccessResult(parsedResult.value as UserFeedSubscriptionId);
-}
-
-/**
  * Creates a new random {@link UserFeedSubscriptionId}.
  */
 export function makeUserFeedSubscriptionId(): UserFeedSubscriptionId {
   return makeUuid<UserFeedSubscriptionId>();
-}
-
-/**
- * Parses a {@link UserFeedSubscription} from an unknown value. Returns an `ErrorResult` if the
- * value is not valid.
- */
-export function parseUserFeedSubscription(
-  maybeUserFeedSubscription: unknown
-): Result<UserFeedSubscription> {
-  const parsedResult = parseZodResult(UserFeedSubscriptionSchema, maybeUserFeedSubscription);
-  if (!parsedResult.success) {
-    return prefixErrorResult(parsedResult, 'Invalid user feed subscription');
-  }
-
-  const parsedUserIdResult = parseUserId(parsedResult.value.userId);
-  if (!parsedUserIdResult.success) return parsedUserIdResult;
-
-  const parsedFeedSourceIdResult = parseFeedSourceId(parsedResult.value.feedSourceId);
-  if (!parsedFeedSourceIdResult.success) return parsedFeedSourceIdResult;
-
-  const parsedUserFeedSubscriptionIdResult = parseUserFeedSubscriptionId(
-    parsedResult.value.userFeedSubscriptionId
-  );
-  if (!parsedUserFeedSubscriptionIdResult.success) return parsedUserFeedSubscriptionIdResult;
-
-  const {url, title, isActive, unsubscribedTime, createdTime, lastUpdatedTime} = parsedResult.value;
-  return makeSuccessResult({
-    userFeedSubscriptionId: parsedUserFeedSubscriptionIdResult.value,
-    feedSourceId: parsedFeedSourceIdResult.value,
-    userId: parsedUserIdResult.value,
-    url,
-    title,
-    isActive,
-    unsubscribedTime: unsubscribedTime
-      ? (new Date(unsubscribedTime) as unknown as Timestamp)
-      : undefined,
-    createdTime: new Date(createdTime) as unknown as Timestamp,
-    lastUpdatedTime: new Date(lastUpdatedTime) as unknown as Timestamp,
-  });
 }
 
 /**
